last homework/requirements.txt
pandas
sqlalchemy
pymysql
watchdog
python-dotenv
